+++ packages/itwin/tree-widget/src/tree-widget-react/components/trees/categories-tree/internal/CategoriesTreeVisibilityHandler.ts
@@ -46,12 +46,12 @@ import {
   getVisibilityFromAlwaysAndNeverDrawnElementsImpl,
   mergeVisibilityStatuses,
 } from "../../common/internal/VisibilityUtils.js";
-import { AlwaysAndNeverDrawnElementInfo } from "../../common/internal/WithoutParents/AlwaysAndNeverDrawnElementInfo.js";
+import { AlwaysAndNeverDrawnElementInfo } from "../../common/internal/withoutParents/AlwaysAndNeverDrawnElementInfo.js";
 import { createVisibilityHandlerResult } from "../../common/UseHierarchyVisibility.js";
 import { CategoriesTreeNode } from "./CategoriesTreeNode.js";
 import { createFilteredTree, parseCategoryKey, parseSubCategoryKey } from "./FilteredTree.js";
 
-import type { CategoryAlwaysOrNeverDrawnElementsQueryProps } from "../../common/internal/WithoutParents/AlwaysAndNeverDrawnElementInfo.js";
+import type { CategoryAlwaysOrNeverDrawnElementsQueryProps } from "../../common/internal/withoutParents/AlwaysAndNeverDrawnElementInfo.js";
 import type { GetVisibilityFromAlwaysAndNeverDrawnElementsProps } from "../../common/internal/VisibilityUtils.js";
 import type { Observable, Subscription } from "rxjs";
 import type { Id64Arg, Id64Array, Id64Set, Id64String } from "@itwin/core-bentley";
